delete user finished

+++ JS/Connections/delete.js
@@ -1,25 +1,30 @@
 //Here is where Delete-request for the profile is gonna be
 async function delete_user(logged_in_user) {
   try {
-    const response = await fetch("../PHP/delete.php", {
-      method: "DELETE",
-      body: JSON.stringify({
-        logged_in_id: logged_in_user.id,
-      }),
-    });
+    const confirmed = confirm("Are you sure you want to delete the user?");
+    if (confirmed) {
+      const response = await fetch("../PHP/delete.php", {
+        method: "DELETE",
+        body: JSON.stringify({
+          logged_in_id: logged_in_user.id,
+        }),
+      });
+      console.log("User deleted");
+      const data = await response.json();
+      console.log(data.message);
 
-    const data = await response.json();
-    console.log(data.message);
-
-    // const photo_container = document.querySelector(
-    //   `img[src="${photo_url}"]`
-    // ).parentNode;
-    // // remove the deleted photo from the UI by finding its parent element and removing it
-    // photo_container.parentNode.removeChild(photo_container);
+      if (response.ok) {
+        // User deleted successfully
+        localStorage.removeItem("user");
+        createHomePage();
+      }
+    }
   } catch (error) {
     console.error(error);
   }
 }
+
+
 
 //With .then
 // function delete_photo(photo_id, photo_url, logged_in_user) {
